minor: pass black

+++ pandora_plugin_arnn/pandora_plugin_arnn.py
@@ -68,7 +68,9 @@ class ARNN(semantic_segmentation.AbstractSemanticSegmentation):
         self._rgb_bands = self.cfg["RGB_bands"]
         self._vegetation_band = self.cfg["vegetation_band"]
 
-    def check_conf(self, **cfg: Dict[str, Union[str, int]]) -> Dict[str, Union[str, int]]:
+    def check_conf(
+        self, **cfg: Dict[str, Union[str, int]]
+    ) -> Dict[str, Union[str, int]]:
         """
         Add default values to the dictionary if there are missing elements and check if the dictionary is correct
 
@@ -103,7 +105,9 @@ class ARNN(semantic_segmentation.AbstractSemanticSegmentation):
         """
         print("Building semantic segmentation")
 
-    def compute_semantic_segmentation(self, cv: xr.Dataset, img_left: xr.Dataset, img_right: xr.Dataset) -> xr.Dataset:
+    def compute_semantic_segmentation(
+        self, cv: xr.Dataset, img_left: xr.Dataset, img_right: xr.Dataset
+    ) -> xr.Dataset:
         """
         Compute building semantic segmentation
 
@@ -132,7 +136,9 @@ class ARNN(semantic_segmentation.AbstractSemanticSegmentation):
         device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
         model = BuildingSegmentation()
         model = model.eval()
-        pretrained_path = files("pandora_plugin_arnn.model").joinpath("pretrained_building_segmentation_model.pt")
+        pretrained_path = files("pandora_plugin_arnn.model").joinpath(
+            "pretrained_building_segmentation_model.pt"
+        )
         model.load_state_dict(torch.load(pretrained_path))
         model = model.to(device)
 
@@ -150,7 +156,9 @@ class ARNN(semantic_segmentation.AbstractSemanticSegmentation):
 
         for indx, band in enumerate(["R", "G", "B"]):
             band_index = list(img_left.band.data).index(self._rgb_bands[band])  # type: ignore
-            rgb_img[indx, begin_row:, begin_col:] = np.copy(img_left["im"].data[band_index, :, :])
+            rgb_img[indx, begin_row:, begin_col:] = np.copy(
+                img_left["im"].data[band_index, :, :]
+            )
 
         model_dataset = xr.Dataset(
             {"im": (["band", "row", "col"], rgb_img.astype(np.float32))},
@@ -184,11 +192,15 @@ class ARNN(semantic_segmentation.AbstractSemanticSegmentation):
             )
 
             # Retrain the network with the annotation map and make a new prediction
-            model_dataset = retrain(model, model_dataset, device, retrain_epoch=1, ignore_index=-1)
+            model_dataset = retrain(
+                model, model_dataset, device, retrain_epoch=1, ignore_index=-1
+            )
 
         # Recrop the segmentation to original image size
         img_left["internal"] = xr.DataArray(
-            data=model_dataset["initial_prediction"].data[begin_row:, begin_col:],
+            data=model_dataset["initial_prediction"].data[
+                begin_row:, begin_col:
+            ],
             coords=[img_left.coords["row"], img_left.coords["col"]],
             dims=["row", "col"],
         )
@@ -237,7 +249,9 @@ class ARNN(semantic_segmentation.AbstractSemanticSegmentation):
         :rtype: xarray.Dataset
         """
         # Apply WTA on cost volume
-        wta = AbstractDisparity(**{"disparity_method": "wta", "invalid_disparity": -9999})
+        wta = AbstractDisparity(
+            **{"disparity_method": "wta", "invalid_disparity": -9999}
+        )
         disp = wta.to_disp(cv, img_left)
         wta.validity_mask(disp, img_left, img_right, cv)
 
@@ -291,9 +305,9 @@ class ARNN(semantic_segmentation.AbstractSemanticSegmentation):
         vegetation_map = self.merge_into_vegetation_map(img_left)
 
         # Dilates the vegetation map and remove vegetation pixel in annotation map
-        vegetation_map = ndimage.binary_dilation(vegetation_map, structure=np.ones((10, 10))).astype(
-            vegetation_map.dtype
-        )
+        vegetation_map = ndimage.binary_dilation(
+            vegetation_map, structure=np.ones((10, 10))
+        ).astype(vegetation_map.dtype)
         annotation[vegetation_map == 1] = -1
 
         return annotation
